+++ nodes/__init__.py
@@ -1,20 +1,2 @@
 # -*- coding: utf-8 -*-
-"""LK Studio Gemini 节点模块"""
-
-from .text_generation import LK_Gemini_Text, LK_Gemini_Chat
-from .image_generation import LK_Gemini_ImageGen, LK_Gemini_ImageEdit
-from .video_generation import LK_Gemini_VideoGen, LK_Gemini_Image2Video
-from .vision_understanding import LK_Gemini_VisionAnalyze, LK_Gemini_DocumentProcess
-from .advanced_features import LK_Gemini_StructuredOutput, LK_Gemini_PromptOptimizer
-from .utility_nodes import LK_Gemini_APIConfig, LK_Gemini_ModelInfo
-from .nano_banana import LK_NanoBanana, LK_NanoBananaPro, LK_ImageToPrompt
-
-__all__ = [
-    'LK_Gemini_Text', 'LK_Gemini_Chat',
-    'LK_Gemini_ImageGen', 'LK_Gemini_ImageEdit',
-    'LK_Gemini_VideoGen', 'LK_Gemini_Image2Video',
-    'LK_Gemini_VisionAnalyze', 'LK_Gemini_DocumentProcess',
-    'LK_Gemini_StructuredOutput', 'LK_Gemini_PromptOptimizer',
-    'LK_Gemini_APIConfig', 'LK_Gemini_ModelInfo',
-    'LK_NanoBanana', 'LK_NanoBananaPro', 'LK_ImageToPrompt'
-]
+"""LK Studio 通用 API 节点模块"""